x11: store dnd.position as pair of i16
It's what translate_coords takes anyway, so the extra precision is
misleading if we're going to cast it to i16 everywhere it's used.

We can also simplify the "unpacking" from the XdndPosition message--we
can and should use the value of 16 as the shift instead of
size_of::<c_short> * 2 or something like that, because the specification
gives us the constant 16.

+++ src/platform_impl/linux/x11/dnd.rs
@@ -46,7 +46,7 @@ pub struct Dnd {
     // Populated by XdndPosition event handler
     pub source_window: Option<xproto::Window>,
     // Populated by XdndPosition event handler
-    pub position: (c_long, c_long),
+    pub position: (i16, i16),
     // Populated by SelectionNotify event handler (triggered by XdndPosition event handler)
     pub result: Option<Result<Vec<PathBuf>, DndDataParseError>>,
     // Populated by SelectionNotify event handler (triggered by XdndPosition event handler)
